requirements.txt
livekit-agents[openai]==0.2.1
python-dotenv==1.0.0
rich==13.5.2
fastapi==0.104.1
uvicorn==0.23.2
openai>=1.3.0 